Assign more reasonable width and height of a window surface
When the surface of a window is created during relayoutWindow,
there is the case that the width and height of the surface are
not given appropriately, e.g. the staring windows.
In this case, the width and height are set to 1, which cause
resizing the size of the surface during setSurfaceBoundariesLocked.
Since setting the size of a surface is processed synchronously
in the SurfaceFlinger, there is usually a few milliseconds
delay (up to 1 vsync interval) when we launch an application.

To remove such cases, this patch assign the width and height
more reasonably like below:

 1. If FLAG_DRAWS_SYSTEM_BAR_BACKGROUNDS is set then assign
    the size of the display we can use.
   (i.e. DisplayInfo.logical[Width|Height])

 2. Otherwise, assign the width and height of the portion of
    the display that is available to applications.
    (i.e. DisplayInfo.app[Width|Height])

Change-Id: Ie8265b9ff0fdb6ecc4577687935adf7cfb4439ad
Signed-off-by: Dohyun Lee <dohyun.lee@lge.com>